requirements-core.txt
# Polymarket & Blockchain (CRITICAL)
py-clob-client==0.28.0
py-order-utils==0.3.2
web3==7.14.0
poly-eip712-structs==0.0.1

# LangChain & AI
langchain==0.3.27
langchain-community==0.3.24
langchain-chroma==0.1.4
openai==2.8.1

# Data & Vector Store
chromadb==0.5.23

# News & Search APIs
newsapi-python==0.2.7
tavily-python==0.7.13
exa-py==2.0.0
kagiapi==0.2.1  # Provides KagiClient class
# perplexityai==0.20.0  # Temporarily disabled - import issues

# Web Framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.36.0,<1.0.0

# HTTP Clients
httpx>=0.28.0,<1.0.0
requests>=2.32.0,<3.0.0

# CLI Tools
typer>=0.19.0,<1.0.0
devtools>=0.12.0,<1.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.12.0,<3.0.0
scheduler==0.8.8